test: refactor check-identifier handler and fix unit tests
