+++ server/api/auth/check-identifier.get.ts
@@ -1,4 +1,5 @@
-export default defineEventHandler(async (event) => {
+import { defineWrappedResponseHandler } from '~~/server/utils/handler';
+export default defineWrappedResponseHandler(async (event) => {
   const query = getQuery<{ identifier: string }>(event);
   const response = await serverApiFetch<ApiSuccessResponse<{ exists: boolean; type: string }>>(
     '/auth/check-identifier',
